fix: potential memory leak when destroying the SDK client (#142)

+++ bucketeer/src/main/kotlin/io/bucketeer/sdk/android/internal/evaluation/db/EvaluationSQLDaoImpl.kt
@@ -2,6 +2,7 @@ package io.bucketeer.sdk.android.internal.evaluation.db
 
 import android.content.ContentValues
 import android.database.sqlite.SQLiteDatabase
+import androidx.annotation.VisibleForTesting
 import androidx.sqlite.db.SupportSQLiteDatabase
 import androidx.sqlite.db.SupportSQLiteOpenHelper
 import com.squareup.moshi.Moshi
@@ -17,13 +18,17 @@ import io.bucketeer.sdk.android.internal.evaluation.db.EvaluationEntity.TABLE_NA
 import io.bucketeer.sdk.android.internal.model.Evaluation
 
 internal class EvaluationSQLDaoImpl(
-  private val sqLiteOpenHelper: SupportSQLiteOpenHelper,
+  @get:VisibleForTesting
+  internal val sqLiteOpenHelper: SupportSQLiteOpenHelper,
   moshi: Moshi,
 ) : EvaluationSQLDao {
-
+  override var isClosed = false
   private val adapter = moshi.adapter(Evaluation::class.java)
 
   override fun put(userId: String, list: List<Evaluation>) {
+    if (isClosed) {
+      return
+    }
     sqLiteOpenHelper.writableDatabase.apply {
       list.forEach { evaluation ->
         val affectedRow = update(this, userId, evaluation)
@@ -41,6 +46,9 @@ internal class EvaluationSQLDaoImpl(
     userId: String,
     evaluation: Evaluation,
   ): Long {
+    if (isClosed) {
+      return -1
+    }
     val contentValue = ContentValues().apply {
       put(COLUMN_USER_ID, userId)
       put(COLUMN_FEATURE_ID, evaluation.featureId)
@@ -54,6 +62,9 @@ internal class EvaluationSQLDaoImpl(
     userId: String,
     evaluation: Evaluation,
   ): Int {
+    if (isClosed) {
+      return 0
+    }
     val contentValues = ContentValues().apply {
       put(COLUMN_EVALUATION, adapter.toJson(evaluation))
     }
@@ -67,6 +78,9 @@ internal class EvaluationSQLDaoImpl(
   }
 
   override fun get(userId: String): List<Evaluation> {
+    if (isClosed) {
+      return listOf()
+    }
     val projection = arrayOf(COLUMN_USER_ID, COLUMN_EVALUATION)
     val c = sqLiteOpenHelper.readableDatabase.select(
       table = TABLE_NAME,
@@ -83,6 +97,9 @@ internal class EvaluationSQLDaoImpl(
   }
 
   override fun deleteAll(userId: String) {
+    if (isClosed) {
+      return
+    }
     sqLiteOpenHelper.writableDatabase.delete(
       TABLE_NAME,
       "$COLUMN_USER_ID=?",
@@ -91,8 +108,17 @@ internal class EvaluationSQLDaoImpl(
   }
 
   override fun startTransaction(block: () -> Unit) {
+    if (isClosed) {
+      return
+    }
     sqLiteOpenHelper.writableDatabase.transaction {
       block()
+    }
+  }
+
+  override fun close() {
+    synchronized(this) {
+      isClosed = true
     }
   }
 }
